:seedling: Add input validations for type 0 transfer transaction
