+++ packages/lisk-transactions/src/0_transfer.js
@@ -12,13 +12,37 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
+import cryptography from 'lisk-cryptography';
 import { BYTESIZES, TRANSFER_FEE } from './constants';
 import {
 	getAddressAndPublicKeyFromRecipientData,
 	wrapTransactionCreator,
+	validateAmount,
 } from './utils';
 
 const createAsset = data => {
+	if (data && data.length > 0) {
+		return { data };
+	}
+	return {};
+};
+
+const validateInputs = ({ amount, recipientId, recipientPublicKey, data }) => {
+	if (!validateAmount(amount)) {
+		// throw new Error('Please enter a valid amount!');
+	}
+
+	if (recipientId && recipientPublicKey) {
+		const addressFromPublicKey = cryptography.getAddressFromPublicKey(
+			recipientPublicKey,
+		);
+		if (recipientId !== addressFromPublicKey) {
+			throw new Error(
+				'Could not create transaction: recipientId does not match recipientPublicKey.',
+			);
+		}
+	}
+
 	if (data && data.length > 0) {
 		if (data.length > BYTESIZES.DATA) {
 			throw new Error('Transaction data field cannot exceed 64 bytes.');
@@ -27,12 +51,11 @@ const createAsset = data => {
 			throw new Error(
 				'Invalid encoding in transaction data. Data must be utf-8 encoded.',
 			);
-		return { data };
 	}
-	return {};
 };
 
 const transfer = ({ amount, recipientId, recipientPublicKey, data }) => {
+	validateInputs({ amount, recipientId, recipientPublicKey, data });
 	const { address, publicKey } = getAddressAndPublicKeyFromRecipientData({
 		recipientId,
 		recipientPublicKey,
