+++ packages/lisk-transactions/src/utils/get_address_and_public_key_from_recipient_data.js
@@ -18,19 +18,7 @@ const getAddressAndPublicKeyFromRecipientData = ({
 	recipientId,
 	recipientPublicKey,
 }) => {
-	if (recipientId && recipientPublicKey) {
-		const addressFromPublicKey = cryptography.getAddressFromPublicKey(
-			recipientPublicKey,
-		);
-		if (recipientId === addressFromPublicKey) {
-			return { address: recipientId, publicKey: recipientPublicKey };
-		}
-		throw new Error(
-			'Could not create transaction: recipientId does not match recipientPublicKey.',
-		);
-	}
-
-	if (!recipientId && recipientPublicKey) {
+	if (recipientPublicKey) {
 		const addressFromPublicKey = cryptography.getAddressFromPublicKey(
 			recipientPublicKey,
 		);
