+++ packages/lisk-transactions/src/utils/validation/validator.js
@@ -18,9 +18,9 @@ import bignum from 'browserify-bignum';
 import {
 	validateAddress,
 	validatePublicKey,
-	isGreaterThanMaxTransactionAmount,
 	isGreaterThanMaxTransactionId,
 	isNumberString,
+	validateAmount,
 } from './validation';
 import * as schemas from './schema';
 
@@ -43,11 +43,7 @@ validator.addFormat('address', data => {
 	}
 });
 
-validator.addFormat(
-	'amount',
-	data =>
-		isNumberString(data) && !isGreaterThanMaxTransactionAmount(bignum(data)),
-);
+validator.addFormat('amount', validateAmount);
 
 validator.addFormat('publicKey', data => {
 	try {
